refactor(backend): extract reconnect notice detection
